% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accPoints.R
\name{accPoints}
\alias{accPoints}
\title{accPoints}
\usage{
accPoints(x, y, xlim, ylim, col, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
Render a simple WebGL-accelerated scatter plot.
Quite useful for large datasets.
}
\details{
Use this function as the first parameter for renderAccPoints.
}
